feat(plenary): move vararg inside async module
